wiregrad/requirements.txt
accelerate==0.30.1
certifi==2024.2.2
charset-normalizer==3.3.2
cholespy==1.0.0
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
contourpy==1.2.1
cycler==0.12.1
diffusers==0.27.2
distro==1.9.0
easydict==1.13
filelock==3.13.1
fonttools==4.51.0
fsspec==2024.2.0
ftfy==6.2.0
huggingface-hub==0.23.0
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
Jinja2==3.1.3
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
libigl==2.5.1
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
nanobind==1.9.2
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
packaging==24.0
pillow==10.2.0
polyscope==2.2.1
psutil==5.9.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
safetensors==0.4.3
scikit-build==0.17.6
scipy==1.13.0
six==1.16.0
sympy==1.12
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0+cu118
torchvision==0.18.0+cu118
tqdm==4.66.4
transformers==4.41.0
triton==2.3.0
typing_extensions==4.9.0
urllib3==2.2.1
wcwidth==0.2.13
zipp==3.18.2
